completed all services implementations, made some changes to types

+++ src/main/java/com/aerotravel/flightticketbooking/services/FlightService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface FlightService {
     public abstract Page<Flight> getAllFlightsPaged(int pageNum);
     public abstract List<Flight> getAllFlights();
-    public abstract Flight getFlightById(Integer flightId);
+    public abstract Flight getFlightById(Long flightId);
     public abstract Flight saveFlight(Flight flight);
     public abstract void deleteFlightById(Long flightId);
 }
